Create profiling.f90
A simple F90 module to keep track of time spent inside each routine of a code. Just import the module and include a call to time_routine at the beginning and the end of the routine. Compatible with MPI parallelization. Call routine time_log to produce output file (or files in case of parallel execution).